+++ plugins/postcss-design-tokens/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Design Tokens
 
+### Unreleased
+
+- Added `functionName` option to control the `design-token` function name.
+
 ### 1.0.0 (May 23, 2022)
 
 - Initial version